backend/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbitrage-backend"
version = "0.1.0"
description = "Backend service for the arbitrage bot prototype"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.1",
    "pydantic-settings==2.5.2",
    "SQLAlchemy==2.0.31",
    "alembic==1.13.1",
    "httpx==0.27.0",
    "redis==5.0.4",
    "PyYAML==6.0.2",
    "hyperliquid-python-sdk==0.20.0"
]

[project.optional-dependencies]
test = [
    "pytest==8.3.2"
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
